function_1/requirements.txt
SQLAlchemy==1.3.12     
PyMySQL==0.9.3
Flask==1.0.2
Click==7.0
google-cloud-storage==1.23.0